feat: impl show image

+++ src/constant.rs
@@ -1,7 +1,7 @@
 // assets/ferris.raw image
 pub const RAW_IMAGE: &[u8] = include_bytes!("../assets/ferris.raw");
-pub const IMAGE_WIDTH: usize = 86;
-pub const IMAGE_HEIGHT: usize = 84;
+pub const IMAGE_WIDTH: u16 = 86;
+pub const IMAGE_HEIGHT: u16 = 84;
 
 // ST7735 display
 pub const ST7735_WIDTH: u16 = 144;
